delete route completed

+++ SQL/index.js
@@ -190,17 +190,55 @@ app.post("/user", (req, res) => {
 // Delete path -
 app.get("/user/:id/delete", (req, res) => {
   let {id} = req.params;
-  let q = `DELETE FROM user WHERE id='${id}'`;
+  let q = `SELECT * FROM user WHERE id='${id}'`;
+  try {
+    connection.query(q, (err, user) => {
+      if(err) throw err;
+      res.render("delete.ejs", {user: user[0]});
+    });
+  } catch(err) {
+    console.log(err);
+    res.send("error in DB");
+  }
+})
+
+// Delete From DB route -
+app.delete("/user/:id", (req, res) => {
+  let {id} = req.params;
+  let {email, password} = req.body;
+  let q = `SELECT * FROM user WHERE id = '${id}'`;
+
   try {
     connection.query(q, (err, result) => {
       if(err) throw err;
-      res.redirect("/user");
+      let user = result[0];
+      if(password != user.password) {
+        res.send("wrong password");
+      } else {
+        let q2 = `DELETE FROM user WHERE id='${id}'`;
+        connection.query(q2, (err, result) => {
+          if(err) throw err;
+          res.redirect("/user");
+        });
+      }
     });
   } catch(err) {
-      console.log(err);
-      res.send("some error in DB");
-    }
+    console.log(err);
+    res.send("error in db");
+  }
 })
+
+
+
+
+
+
+
+
+
+
+
+
 
 app.listen("8080", () => {
   console.log("server is listening on port 8080");
